FIX: When user uses custom electrum server, make him aware that his addresses/txids sre still leaked to a Push Notificatinos serer #4806
