+++ loc/en.json
@@ -295,9 +295,10 @@
     "privacy_clipboard_explanation": "Provide shortcuts if an address or invoice is found in your clipboard.",
     "privacy_do_not_track": "Disable Analytics",
     "privacy_do_not_track_explanation": "Performance and reliability information will not be submitted for analysis.",
-    "push_notifications": "Push Notifications",
     "rate": "Rate",
-    "selfTest": "Self-Test",
+
+    "push_notifications_explanation": "By enabling notifications, the addresses for all wallets in the app and transaction IDs from any transactions made with the app after enabling notifications will be sent to our server.\nThis is necessary so that we can notify you of updates such as incoming Bitcoin or transaction confirmations.\n\nOnly wallet addresses and transaction IDs created after enabling notifications are transmitted—nothing prior is gathered.\n\nDisabling notifications will remove all of this information from the server."
+,    "selfTest": "Self-Test",
     "save": "Save",
     "saved": "Saved",
     "success_transaction_broadcasted": "Your transaction has been successfully broadcasted!",
