+++ navigation/LazyLoadSettingsStack.tsx
@@ -13,7 +13,7 @@ const DefaultView = lazy(() => import('../screen/settings/DefaultView'));
 const ElectrumSettings = lazy(() => import('../screen/settings/electrumSettings'));
 const EncryptStorage = lazy(() => import('../screen/settings/EncryptStorage'));
 const LightningSettings = lazy(() => import('../screen/settings/LightningSettings'));
-const NotificationSettings = lazy(() => import('../screen/settings/notificationSettings'));
+const NotificationSettings = lazy(() => import('../screen/settings/NotificationSettings'));
 const SelfTest = lazy(() => import('../screen/settings/SelfTest'));
 const ReleaseNotes = lazy(() => import('../screen/settings/ReleaseNotes'));
 const Tools = lazy(() => import('../screen/settings/tools'));
